+++ src/wbs/applications/imchat/api/ImChatConversationData.java
@@ -2,7 +2,6 @@ package wbs.applications.imchat.api;
 
 import lombok.Data;
 import lombok.experimental.Accessors;
-
 import wbs.framework.data.annotations.DataAttribute;
 import wbs.framework.data.annotations.DataClass;
 
@@ -13,7 +12,7 @@ public
 class ImChatConversationData {
 
 	@DataAttribute
-	Integer index;
+	Long index;
 
 	@DataAttribute
 	ImChatProfileData profile;
