+++ src/wbs/applications/imchat/api/ImChatMediaOriginalJpegAction.java
@@ -6,7 +6,6 @@ import javax.inject.Inject;
 import javax.inject.Provider;
 
 import lombok.Cleanup;
-
 import wbs.applications.imchat.model.ImChatObjectHelper;
 import wbs.applications.imchat.model.ImChatProfileObjectHelper;
 import wbs.framework.application.annotations.PrototypeComponent;
@@ -69,7 +68,7 @@ class ImChatMediaOriginalJpegAction
 
 		MediaRec media =
 			mediaHelper.findOrThrow (
-				Integer.parseInt (
+				Long.parseLong (
 					requestContext.requestStringRequired (
 						"mediaId")),
 				() -> new PageNotFoundException ());
@@ -79,7 +78,7 @@ class ImChatMediaOriginalJpegAction
 		ContentRec content =
 			media.getContent ();
 
-		Integer hash =
+		Long hash =
 			Math.abs (
 				content.getHash ());
